fix(10-02): fix all type mismatch and sign comparison warnings
- Change int loop vars to uint32_t/size_t to match struct field types
- Add parentheses for -Wparentheses in mixed &&/|| and &/| expressions
- Fix int64_t format specifiers: %lld -> PRId64 for portability
- Remove unused variables (all_positive, both_infeasible, count0/count1, etc)
- Add const qualifiers to get_index parameters
- Use size_t for max_item comparisons in eval_constraint
- Zero warnings with -Wall -Wextra -Wpedantic -Wvla -Werror
- All 56 tests pass

+++ cbqs/src/state.c
@@ -1,5 +1,6 @@
 #include "state.h"
 #include <math.h>
+#include <inttypes.h>
 
 int min(int a, int b){
     return (a < b) ? a : b;
@@ -79,7 +80,7 @@ void copy_state_inplace(state_t *dest, state_t *src){
 }
 
 void print_state(state_t *state){
-    printf("%lld %f %d %zu ", state->tot_profit, state->prob, state->feasible, state->vector.bits);
+    printf("%" PRId64 " %f %d %zu ", state->tot_profit, state->prob, state->feasible, state->vector.bits);
     sw_print(state->vector);
     printf(" ");
     sw_print(state->branch);
@@ -143,7 +144,7 @@ state_t *read_states(char **name, int num_files, size_t *NumberStatesFinal, int 
     }
 //    printf("count %d\n", count);
     *NumberStatesFinal = count;
-    for (int i = count; i < estimate; i++){
+    for (size_t i = count; i < estimate; i++){
         sw_clear(parent[i].vector);
         sw_clear(parent[i].branch);
     }
